jinho1011/ministagram/requirements.txt
aiofile==1.5.2
aiohttp==3.6.2
alabaster==0.7.12
altgraph==0.17
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
autopep8==1.5.2
Babel==2.6.0
beautiful-date==1.0.1
beautifulsoup4==4.9.1
bs4==0.0.1
cached-property==1.5.1
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
click==7.1.2
commonmark==0.9.1
dictdiffer==0.8.1
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
docutils==0.14
flake8==3.8.3
Flask==1.1.2
future==0.18.2
gcsa==0.2.1
google-api-core==1.21.0
google-api-python-client==1.7.7
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
googleapis-common-protos==1.52.0
httplib2==0.18.0
idna==2.8
imagesize==1.1.0
importlib-metadata==1.6.1
install==1.3.3
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.5.1
lzstring==1.0.4
macholib==1.14
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==8.4.0
multidict==4.7.6
notion==0.0.25
oauth2client==4.1.3
oauthlib==3.0.1
packaging==19.0
pluggy==0.13.1
protobuf==3.12.2
py==1.8.1
py2exe==0.9.2.2
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
PyInstaller==3.6
pyparsing==2.3.1
pyseoanalyzer==4.0.2
pytest==5.4.3
python-dateutil==2.7.2
python-slugify==4.0.0
pytz==2018.9
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
screen==1.0.1
selenium==3.141.0
sitemap-generator==0.9.2
six==1.11.0
snowballstemmer==1.2.1
soupsieve==2.0.1
Sphinx==1.8.3
sphinxcontrib-websupport==1.1.0
text-unidecode==1.3
tzlocal==1.5.1
ultimate-sitemap-parser==0.5
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.2.4
Werkzeug==1.0.1
xlwt==1.3.0
xmltodict==0.12.0
yarl==1.4.2
zipp==3.1.0
